+++ src/extension.ts
@@ -5,12 +5,12 @@ import { Folders } from './commands/Folders';
 import { Preview } from './commands/Preview';
 import { Project } from './commands/Project';
 import { Template } from './commands/Template';
-import { COMMAND_NAME, EXTENSION_ID, EXTENSION_STATE_VERSION } from './constants/Extension';
+import { COMMAND_NAME } from './constants/Extension';
 import { TaxonomyType } from './models';
 import { MarkdownFoldingProvider } from './providers/MarkdownFoldingProvider';
 import { TagType } from './viewpanel/TagType';
 import { ExplorerView } from './webview/ExplorerView';
-import { Notifications } from './helpers/Notifications';
+import { Extension } from './helpers/Extension';
 
 let frontMatterStatusBar: vscode.StatusBarItem;
 let statusDebouncer: { (fnc: any, time: number): void; };
@@ -20,11 +20,9 @@ let collection: vscode.DiagnosticCollection;
 const mdSelector: vscode.DocumentSelector = { language: 'markdown', scheme: 'file' };
 
 export async function activate({ subscriptions, extensionUri, extensionPath, globalState }: vscode.ExtensionContext) {
-	collection = vscode.languages.createDiagnosticCollection('frontMatter');
+	const extension = Extension.getInstance(globalState, extensionPath);
 
-	const frontMatter = vscode.extensions.getExtension(EXTENSION_ID)!;
-	const frontMatterVersoin = frontMatter.packageJSON.version;
-	const crntVersion = globalState.get<string>(EXTENSION_STATE_VERSION);
+	collection = vscode.languages.createDiagnosticCollection('frontMatter');
 
 	// Pages dashboard
 	Dashboard.init(extensionPath);
@@ -32,7 +30,7 @@ export async function activate({ subscriptions, extensionUri, extensionPath, glo
 		Dashboard.open(extensionPath);
 	}));
 
-	if (!crntVersion) {
+	if (!extension.getVersion().usedVersion) {
 		vscode.commands.executeCommand(COMMAND_NAME.dashboard);
 	}