+++ src/models/VersionInfo.ts
@@ -1,0 +1,6 @@
+
+
+export interface VersionInfo { 
+  usedVersion: string | undefined;
+  installedVersion: string;
+}