+++ src/pagesView/DashboardMessage.ts
@@ -4,4 +4,6 @@ export enum DashboardMessage {
   getTheme = 'getTheme',
   createContent = 'createContent',
   updateSetting = 'updateSetting',
+  InitializeProject = 'InitializeProject',
+  Reload = 'Reload',
 }