+++ src/pagesView/components/Dashboard.tsx
@@ -7,10 +7,14 @@ import { Tab } from '../constants/Tab';
 import { SortOption } from '../constants/SortOption';
 import useDarkMode from '../../hooks/useDarkMode';
 import usePages from '../hooks/usePages';
+import { SponsorMsg } from './SponsorMsg';
+import { WelcomeScreen } from './WelcomeScreen';
 
-export interface IDashboardProps {}
+export interface IDashboardProps {
+  showWelcome: boolean;
+}
 
-export const Dashboard: React.FunctionComponent<IDashboardProps> = ({}: React.PropsWithChildren<IDashboardProps>) => {
+export const Dashboard: React.FunctionComponent<IDashboardProps> = ({showWelcome}: React.PropsWithChildren<IDashboardProps>) => {
   const { loading, pages, settings } = useMessages();
   const [ tab, setTab ] = React.useState(Tab.All);
   const [ sorting, setSorting ] = React.useState(SortOption.LastModified);
@@ -22,6 +26,18 @@ export const Dashboard: React.FunctionComponent<IDashboardProps> = ({}: React.Pr
   useDarkMode();
 
   const pageGroups = [...new Set(pages.map(page => page.fmGroup))];
+
+  if (!settings) {
+    return <Spinner />;
+  }
+
+  if (showWelcome) {
+    return <WelcomeScreen settings={settings} />;
+  }
+
+  if (!settings.initialized || settings.folders?.length === 0) {
+    return <WelcomeScreen settings={settings} />;
+  }
   
   return (
     <main className={`h-full w-full`}>
@@ -46,7 +62,7 @@ export const Dashboard: React.FunctionComponent<IDashboardProps> = ({}: React.Pr
           { loading ? <Spinner /> : <Overview pages={pageItems} settings={settings} /> }
         </div>
 
-        <p className={`text-vulcan-50 dark:text-whisper-900 py-2 text-center`}>If you find FrontMatter useful, please consider <a className={`text-vulcan-500 dark:text-whisper-500 hover:text-vulcan-600 dark:hover:text-whisper-300`} href="https://github.com/sponsors/estruyf" title="Sponsor Front Matter">sponsoring</a> it, or if you use FrontMatter for work, please encourage your employer to <a className={`text-vulcan-500 dark:text-whisper-500 hover:text-vulcan-600 dark:hover:text-whisper-300`} href="https://github.com/sponsors/estruyf" title="Sponsor Front Matter">sponsoring</a> it.</p>
+        <SponsorMsg />
       </div>
     </main>
   );