+++ src/pagesView/models/Status.ts
@@ -1,0 +1,7 @@
+
+
+export enum Status {
+  Active = 1,
+  Completed,
+  NotStarted
+}